added sales report and invoice

+++ models/product.go
@@ -1,14 +1,14 @@
 package models
 
 type Product struct {
-	ID          int    `json:"id"`
-	Name        string `json:"name" `
-	Description string `json:"description"`
-	Category    string `json:"category"`
-	Size        string `json:"size"`
-	Stock       int    `json:"stock"`
-	Price       int    `json:"price"`
-	Color       string `json:"color"`
+	ID          int     `json:"id"`
+	Name        string  `json:"name" `
+	Description string  `json:"description"`
+	Category    string  `json:"category"`
+	Size        string  `json:"size"`
+	Stock       int     `json:"stock"`
+	Price       float64 `json:"price"`
+	Color       string  `json:"color"`
 }
 
 type UpdateProduct struct {
